separate full and light node RPC
