+++ client/Cargo.toml
@@ -43,6 +43,7 @@ tempdir = "0.3"
 rustc-hex= "1.0"
 threadpool = "1.0"
 metrics = { path = "../util/metrics" }
+delegate = "0.2.0"
 
 [dev-dependencies]
 criterion = "0.2"
