+++ client/src/rpc/impls.rs
@@ -3,3 +3,5 @@
 // See http://www.gnu.org/licenses/
 
 pub mod cfx;
+pub mod common;
+pub mod light;
